+++ server/lib/product-variants/index.js
@@ -2,8 +2,10 @@
 
 const schema = require('./schema');
 const service = require('./service');
+const colorHex = require('./color-hex');
 
 module.exports = {
   ...schema,
   ...service,
+  ...colorHex,
 };